czfef

+++ backCake.php
@@ -105,7 +105,7 @@ if(!isset($_SESSION['verif'])){
                             <?php echo $newValue['description'] ?> 
                             <!-- <input type="hidden" value="<?php$value['gateau_id']?>"> -->
     
-                            <a href="delete.php?id=<?php echo $newValue['id'];?>">supprimer</a>
+                            <a href="delete.php?id=<?php echo $newValue['id'];?>" onclick= "return confirm('etes vous sure?')">supprimer</a>
                             <a href="modif.php?id=<?php echo $newValue['id'];?>">Modifier</a>
     
     
@@ -134,7 +134,7 @@ if(!isset($_SESSION['verif'])){
                     <?php echo $newCle['description'] ?> 
                     <!-- <input type="hidden" value="<?php$newCle['gateau_id']?>"> -->
     
-                    <a href="delete.php?id=<?php echo $newCle['id'];?>" class="delete">supprimer</a>
+                    <a href="delete.php?id=<?php echo $newCle['id'];?>" onclick= "return confirm('etes vous sure?')">supprimer</a>
                     <a href="modif.php?id=<?php echo $newCle['id'];?>">Modifier</a>
     
     
@@ -162,5 +162,5 @@ if(!isset($_SESSION['verif'])){
 </body>
 <script src="./gateau.js" defer></script>
 <script src="./uploadFichier.js" defer></script>
-<script src="./sure.js" defer></script>
+
 </html>
